[feat]: add enum ErrorHandling
에러 처리해주는 열거형 ErrorHandling 생성
