+++ JuiceMaker/JuiceMaker/Model/ErrorHandling.swift
@@ -1,0 +1,13 @@
+//
+//  ErrorHandling.swift
+//  JuiceMaker
+//
+//  Created by Marisol Mirim Kim on 2022/02/22.
+//
+
+import Foundation
+
+enum ErrorHandling: Error {
+    case outOfStock
+    case wrongFormat
+}
