+++ JuiceMaker/JuiceMaker/Model/FruitStore.swift
@@ -9,9 +9,9 @@ import Foundation
 // 과일 저장소 타입
 class FruitStore {
     static var fruitList: Dictionary = [FruitType.strawberry: 10,
-                                 FruitType.banana: 10,
-                                 FruitType.pineapple: 10,
-                                 FruitType.kiwi: 10,
-                                 FruitType.mango: 10]
+                                        FruitType.banana: 10,
+                                        FruitType.pineapple: 10,
+                                        FruitType.kiwi: 10,
+                                        FruitType.mango: 10]
 }
